requirements.txt
cffi==1.14.4
click==7.1.2
gphoto2==2.2.4
prompt-toolkit==3.0.8
pycairo==1.20.0
pycparser==2.20
PyMTP==0.0.6
pyusb==1.1.0
questionary==1.8.1
wcwidth==0.2.5
